requirements.txt
pytest==9.0.2
selenium==4.18.1
requests==2.31.0
allure-pytest==2.13.2
python-dotenv==1.0.1
webdriver-manager==4.0.1